requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.10
django-phonenumber-field==6.0.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
packaging==21.3
phonenumbers==8.12.40
phonenumberslite==8.12.40
Pillow==8.4.0
pkg_resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
redis==4.1.0
redis-server==6.0.9
requests==2.26.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
twilio==7.4.0
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
zipp==3.6.0
zope.interface==5.4.0
